+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Symlinked site configuration file is not taken into account
+
 ## [3.2.0] - 2025-06-15
 
 ### Added